+++ Runtime/Scripts/IDeferAgent.cs
@@ -24,13 +24,33 @@ namespace GLTFast {
     /// </summary>
     public interface IDeferAgent {
         /// <summary>
-        /// This will be called by GltFast at various points in the procedure.
+        /// This will be called by GltFast at various points in the loading procedure.
         /// </summary>
         /// <returns>True if the remaining work of the loading procedure should
         /// be deferred to the next frame/Update loop invocation. False if
         /// work can continue.</returns>
         bool ShouldDefer();
 
+        /// <summary>
+        /// Indicates if upcoming work should be deferred to the next frame.
+        /// </summary>
+        /// <param name="duration">Predicted duration of upcoming processing in seconds</param>
+        /// <returns>True if the remaining work of the loading procedure should
+        /// be deferred to the next frame/Update loop invocation. False if
+        /// work can continue.</returns>
+        bool ShouldDefer( float duration );
+
+        /// <summary>
+        /// Conditional yield. May continue right away or yield once, based on time.
+        /// </summary>
+        /// <returns>If <see cref="ShoudDefer"/> returns true, returns Task.Yield(). Otherwise returns sync</returns>
         Task BreakPoint();
+        
+        /// <summary>
+        /// Conditional yield. May continue right away or yield once, based on time and duration.
+        /// </summary>
+        /// <param name="duration">Predicted duration of upcoming processing in seconds</param>
+        /// <returns>If <see cref="ShoudDefer"/> returns true, returns Task.Yield(). Otherwise returns sync</returns>
+        Task BreakPoint( float duration );
     }
 }